requirements.txt
# ============================================================
# AI Glass Cleanliness Detection System
# Requirements
# ============================================================

# --- Core Deep Learning ---
torch>=2.1.0
torchvision>=0.16.0

# --- Computer Vision ---
opencv-python>=4.8.0
Pillow>=10.0.0

# --- Data Augmentation ---
albumentations>=1.3.1

# --- Model Libraries ---
segmentation-models-pytorch>=0.3.3
timm>=0.9.12
einops>=0.7.0

# --- Numerics & Science ---
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.2

# --- Config & Utilities ---
pyyaml>=6.0.1
tqdm>=4.66.0
python-dotenv>=1.0.0

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0
tensorboard>=2.15.0
tensorboardX>=2.6.2

# --- Video Processing ---
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# --- Logging & Experiment Tracking ---
wandb>=0.16.0

# --- Development & Testing ---
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0

# --- Notebook Exploration ---
jupyter>=1.0.0
ipywidgets>=8.1.0

# --- Type Checking ---
mypy>=1.7.0
